data: use int64 for less type conversion

+++ internal/data/sample.go
@@ -22,7 +22,7 @@ type Sample struct {
 }
 
 // NewSample creates a new data sample with the given arguments.
-func NewSample(index, size int) Sample {
+func NewSample(index, size int64) Sample {
 	return Sample{
 		ItemBase: btree.NewItemBase(index, size),
 	}
